misc/HtmlTestSuite/results/test-0820.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Animations: parsing animation-name with valid values</title>
<link href="https://drafts.csswg.org/css-animations/#propdef-animation-name" rel="help"/>
<meta content="animation-name supports the full grammar '[ none | &lt;keyframes-name&gt; ]#'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("animation-name", 'NONE', 'none');

test_valid_value("animation-name", 'foo');
test_valid_value("animation-name", 'Both');
test_valid_value("animation-name", 'ease-in');
test_valid_value("animation-name", 'infinite');
test_valid_value("animation-name", 'paused');
test_valid_value("animation-name", 'first, second, third');

test_valid_value("animation-name", '"string"', ['"string"', "string"]);
test_valid_value("animation-name", '"multi word string"', ['"multi word string"', "multi\\ word\\ string"]);
test_valid_value("animation-name", '"initial"');
test_valid_value("animation-name", '"---\\22---"', ['\"---\\\"---\"', '---\\\"---']);
</script>
</body>
</html>
